docker git clone, auto update

+++ docker/source/update.sh
@@ -1,19 +1,23 @@
 #!/bin/bash
 
-TOKEN="3_ccrwj9fQrX8hcFzN95"
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+if [ "$GIT_TOKEN" == "" ]; then
+    echo "No token provided, skipping clone and build step"
+else
+    DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
-cd $DIR
-if [ -d shine-backend ]; then
-    cd shine-backend
-    git reset --hard
-    git checkout master
-    git pull
-else     
-    git clone https://oauth2:${TOKEN}@gitlab.com/gzp/shine-backend.git
-    cd shine-backend
+    cd $DIR
+    if [ -d shine-backend ]; then
+        cd shine-backend
+        git reset --hard
+        git checkout master
+        git pull
+    else     
+        git clone https://oauth2:${GIT_TOKEN}@gitlab.com/gzp/shine-backend.git
+        cd shine-backend
+    fi
+
+    rm -f $CARGO_TARGET_DIR/release/shine-auth
+    cargo build --release
+
+    cp $CARGO_TARGET_DIR/release/shine-auth /webapp/binary
 fi
-
-cargo build --release
-
-cp $CARGO_TARGET_DIR/release/shine-auth /webapp/binary
